Datastore set up, verify repository method working

+++ app/src/main/java/com/codepunk/hollarhype/data/local/dao/UserDao.kt
@@ -19,7 +19,7 @@ abstract class UserDao {
     abstract suspend fun insertUsers(users: List<LocalUser>)
 
     @Query("SELECT * FROM user WHERE id = :userId")
-    abstract fun getUser(userId: Int): Flow<LocalUser?>
+    abstract fun getUser(userId: Long): Flow<LocalUser?>
 
     // endregion Methods
